fix: gitAtime build fail on 32-bit FreeBSD and NetBSD
(cherry picked from commit b94799068be211e4df92ed78ebdf36cec582805a)

+++ interp/os_atimespec.go
@@ -13,5 +13,6 @@ func getAtime(info fs.FileInfo) time.Time {
 	if !ok {
 		return info.ModTime()
 	}
-	return time.Unix(stat.Atimespec.Sec, stat.Atimespec.Nsec)
+	// Note that these are int32 on some platforms.
+	return time.Unix(int64(stat.Atimespec.Sec), int64(stat.Atimespec.Nsec))
 }
